+++ packages/salesforcedx-apex-replay-debugger/src/constants.ts
@@ -4,6 +4,8 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+
+export const ADDRESS_PREFIX = '0x';
 export const CHECKPOINT = 'checkpoint';
 export const CHECKPOINTS_LOCK_STRING = 'checkpoints_lock';
 export const DEBUGGER_LAUNCH_TYPE = 'launch';
