+++ packages/salesforcedx-apex-replay-debugger/src/messages/i18n.ts
@@ -34,5 +34,5 @@ export const messages = {
   malformed_log_line:
     "Encountered a malformed HEAP_DUMP log line, skipping. Log line number=%d, log line='%s'.",
   reconcile_heapdump_error:
-    'Problems were encountered while using heapdump: %s.'
+    'Problems were encountered while using a heap dump: %s. Use Tooling API to delete the ApexExecutionOverlayResult record with the ID %s: From a terminal, run "sfdx force:data:record:delete -t -s ApexExecutionOverlayResult -i %s".'
 };
